requirements.txt
PyMySQL~=1.1.1
python-dotenv==1.0.1